Create Function Alternative to a Nested Loop
Embedded loops can be confusing to write and confusing to read.  You might also find out that the code in the embedded loop is even more useful than you realized, and that you actually need it somewhere else in your code.  What are you going to do? Copy and paste it?  That sounds risky.  What if you make a copy/paste error? What if you need to update the code?  What will you do then? Update both copies?  What if you don't remember where you copied it to or from?

For example, consider this program.  There are two loops: The inner loop calculates the total of all test scores for a student. The outer loop takes the inner loop's total and uses it to calculate the average of all test scores for each student.  

#include <iostream>
using namespace std;

int main() {
   int numstudents = 2;
   int numTests = 3;
   int total;
   int average;
  
   //Determine each student's average score
   for(int student = 1; student <= numstudents; student++)
   {
      total = 0;
      for (int test = 1; test <=numTests; test++)
      {
         double score;
         cout << "Enter score " << test << " for ";
         cout << " student " << student << " : ";
         cin >> score;
         total += score;
      }
      average = total/numTests;
      cout << "The average score for student " << student;
      cout << " is " << average << ".\n\n";
   }
 
}
What if we created a separate function to calculate the total of all tests for a student (the inner loop)?  We could call the function anytime we need it: inside the loop that uses it to calculate the average test score, and also elsewhere in the program.  It would be easier to read, easier to test, easier to reuse and easier to maintain.  To do this, we need to consider which data is needed in the function and how to get it there. To calculate the total of a student's tests, we need a few variables: numTests, student and total.  What are they for and where should they come from?

numTests: This will store the number of tests.  It is used in the inner loop condition for continuing or stopping the loop.  It could come from the user in the form of keyboard input or from a file.  To keep things simple and easier to reuse, we will let the main program decide where to get the value for numTests and the main function will pass it on to the function.  So, this will be a parameter in our new function.
student:  It is used to control the outer loop in the original program, but we need to display it in our new function. This must come from the outer loop in main, so it will also be a parameter.
total: this is the result of the loop calculation in our new function.  Because it is 0 to start, it could come from main as 0, or it could be declared in the new function as 0. Either way would work, so this might be a parameter as well. The main function will need to get the total once the new function is done, so it might also be a return value.
So now, which kind of parameters should we use?  We have two options:

Option 1: pass-by-value parameters (we will need a return value)
Option 2: pass-by-value parameters and a pass-by-reference parameter (instead of a return value)
Let's explore each option.

Option 1: pass-by-value parameters and a return value