Create small multiplication table:
It is possible to embed one loop of any type into another loop of any type.

A nested loop is a loop inside the body of another loop​
The Inner loop is inside the curly braces of the outer (outside) loop.​
​
Here is a simple example that displays a small multiplication table:

for (row=1; row<=3; row++)  //outer​
{
   for (col=1; col<=3; col++)//inner​
   {
        cout << row * col << endl;
    }
 }
  
The next example will calculate the average of each students' test scores.  There are two loops: The inner loop calculates the total of all test scores for a student. The outer loop takes the inner loop's total and uses it to calculate the average of all test scores for each student.  

#include <iostream>
using namespace std;

int main() {
   int numstudents = 2;
   int numTests = 3;
   int total;
   int average;
  
   //Determine each student's average score
   for(int student = 1; student <= numstudents; student++)
   {
      total = 0;
      for (int test = 1; test <=numTests; test++)
      {
         double score;
         cout << "Enter score " << test << " for ";
         cout << " student " << student << " : ";
         cin >> score;
         total += score;
      }
      average = total/numTests;
      cout << "The average score for student " << student;
      cout << " is " << average << ".\n\n";
   }
 
}
Can you see the inner and outer loops?