Create  outer loop takes the inner loop's total and uses it to calculate
#include <iostream>
using namespace std;

int main() {
   int numstudents = 2;
   int numTests = 3;
   int total;
   int average;
  
   //Determine each student's average score
   for(int student = 1; student <= numstudents; student++)
   {
      total = 0;
      for (int test = 1; test <=numTests; test++)
      {
         double score;
         cout << "Enter score " << test << " for ";
         cout << " student " << student << " : ";
         cin >> score;
         total += score;
      }
      average = total/numTests;
      cout << "The average score for student " << student;
      cout << " is " << average << ".\n\n";
   }
 
}
Can you see the inner and outer loops?

nested loop

Here are a few things to note about nested loops:

Any type of loop can be used for the inner and outer loops.  The inner loop does not have to be the same type of loop as the outer loop.
Inner loop goes through all repetitions for each repetition of outer loop​​
Inner loop repetitions complete sooner than outer loop
Total number of repetitions for inner loop is product of number of repetitions of the two loops.  
Variables declared in the outer loop are accessible by the inner loop.  In the above example, student is declared in the outer for loop and can be used to display the student in the inner loop.
Variables declared in the inner loop are not accessible by the outer loop.​  In the above example, test is declared in the inner for loop.  Trying to modify or access the value of the test variable in the outer loop would produce an error.