Create  pass-by-reference parameter (instead of a return value)
For this option, we do not need a return value, so we will create a void function.  Since the main function needs the total, we will use a reference parameter for that.

void calculateTotal(int numTests, int student, int& total)
{
}
No return type is needed, because the total variable will be updated in the main function automatically, thanks to the & (yay, reference parameter!). So, all we need is the loop to calculate the student's total:

void calculateTotal(int numTests, int student, int& total)
{
   for (int test = 1; test <=numTests; test++)
  {
      double score;
      cout << "Enter score " << test << " for ";
      cout << " student " << student << " : ";
      cin >> score;
      total += score;
  }
}
To call this function in main, we do not need to catch any return values, because it is a void function that does not return anything. We just need to send over all of the arguments, including a variable for the reference parameter called total. We will make sure it is set to 0 when it is declared:

int main() {
   int numstudents = 2;
   int numTests = 3;
   
   int average;
  //Determine each student's average score
   
   for(int student = 1; student <= numstudents; student++)
   {
      cout << "Calling by reference: " << endl;
      calculateTotal(numTests, student, total); 
      average = total/numTests;
      cout << "The average score for student " << student;
      cout << " is " << average << ".\n\n";
   }
}
The problem with this example is that the total variable is acting as an accumulator variable for the loop of all students.  To make the accumulator variable reset for each student, move the declaration of the total variable to the body of the loop.  Here is the complete program:

#include <iostream>
using namespace std;

void calculateTotal(int numTests, int student, int& total)
{
   for (int test = 1; test <=numTests; test++)
  {
      double score;
      cout << "Enter score " << test << " for ";
      cout << " student " << student << " : ";
      cin >> score;
      total += score;
  }
}
int main() {
   int numstudents = 2;
   int numTests = 3;
   int average;
  //Determine each student's average score
   
   for(int student = 1; student <= numstudents; student++)
   {
      int total = 0; //reset total for each student
      cout << "Calling by reference: " << endl;
      calculateTotal(numTests, student, total); 
      average = total/numTests;
      cout << "The average score for student " << student;
      cout << " is " << average << ".\n\n";
   }
}