Create hheelloooolleehh
Right to Left
If you need to access each character in a string from right to left (backwards), your for loop header would look like this:

for(int i = name_of_string.length() - 1; i>=0; i--)
Let's look at each section of the code in parentheses:

int i = name_of_string.length() - 1;
This sets i equal to the last valid index in the string (the right-most character).
i>=0;
This will keep the loop going until i is equal to 0, which is the first index in the string (left-most character)
i--
This makes the index value smaller with each iteration of the loop, so it moves from right to left (backward) through the string.
Let's go back to that first point. Notice that the for loop begins at name_of_string.length() - 1. Why? First, let's look at the lengthLinks to an external site. function a little closer.  The reference page says that the length function "returns the length of the string, in terms of bytes."  When the string is in English, the length of the string is the number or characters in the string.  But there is a catch: the characters are counted beginning with the number 1.  So, a string with only one character has a length of 1.  Remember that the index numbers for each character begin with 0, so the index number of the first character is 0.  

Now, let's look at a string

ABCDEFGH

If you count the letters, you can see that there are 8 characters, but the index numbers go from 0 to 7.  So, if you use the length function to get the number of characters in the string, you will get a number (8) that is one larger than the highest valid index number (7) in the string.  Since we want to start the loop at the last character's index number, we need to start at 7.  So, we subtract 1 from the length of the string to start our loop.

ABCDEFGH with length 9 and last valid index 7

Each Character In Between


So, now we have a for loop header that will iterate from left-to-right across a string and another header that will iterate from right-to-left across a string. But what do we do with each character in between?  The key lies with the i variable.  That counter variable called i represents each index number in the string.  So, to access each and every individual character, use the variable i in the square brackets, or with the at() function like this example that displays each character in i using [ ] and a space, then it displays each character in i again using at().