Create cipher
Need a few hints? Let's look at the similarities between the secret and the decode functions first. 

These functions each have two parameters:

string encode/decode: These are the strings that will be used to create another string
int key: This is the integer key that will be used with the modulo operator in the function
The functions have three local variables:

string encoded/decoded: These variables will hold the modified string
string test: This variable will hold a single character of the modified string
int cipher: This variable will hold the result of the key parameter value after it has been limited by the modulo operator.  
The functions have a loop that begins at the beginning of the encode/decode string and execute the body of the loop for each character from left to right:

 for(int i = 0; i < encode.length(); i++), 

The loop calculates the value for the cipher variable with the modulo operator, which limits the possible values to a value between 0 - 25.  See the module topic The Modulo Operator % .  

Now, let's look at the difference between these functions: the modulo operation:

The loop sets the test variable equal to the current letter in the loop plus or minus the cipher integer value. 

Wait, what? You can add a number to a character? 

Actually, yes.  Characters have an integer value assigned by the character set it belongs to.  Here is the ASCII encoding table:

https://www.ascii-code.com/Links to an external site.

The column labeled Dec on that website represents the number assigned to each character listed in the column labeled Symbol.  For example, Capital letters are assigned an ASCII value between 65 -90 and lower case letters have a value between 97-122.  So, when you add a number to an ASCII value, you get a character that is further down on the ASCII table.  If you subtract a number from an ASCII value, you get a character that is closer to the top of the ASCII table.

Did you notice that the secret function adds a number to the character, but the decode  function subtracts that same number from a character?  What would that do to the final strings that are returned by each function?

If you haven't already figured it out, this program will encode a string as a secret message using a key, and the decode function will decode the secret message back to its original plain text version using that same key.  It is a very simple encryption program.