Create The Swap Algorithm + forLoop
So you want to swap two letters in a string?  Let's look at how to swap items from scratch.

I will put the letters of a string "ABCDEFGH" in individual boxes.  Now our string looks as follows:

ABCDEFGH our example string

Now, I will number each box, beginning with 0.  These numbers are called index numbers and they always start with 0.  Each index number indicated the relative position of each letter in the string.  Now, we will swap the item at the (arbitrarily chosen) second index position with the item in the (also arbitrarily chosen) fifth index position:

Getting ready to swap C and F in string ABCDEFGH

First, temporarily store the contents of one of the items we want to swap. For simplicity, let's just choose the first item as the thing we want to store...temporarily:

Copy of C kept separately from string

With the contents of the first item we want to swap temporarily stored, the next step is to assign the contents of the second item to the first item:

swaparray4.png

Your string is in an awkward state right now with a duplicate letter: both the locations you want to swap store the same contents. This is only temporary.  The next step is to take the temporary contents you stored earlier from the first item and assign it to our second item:

Finalize swap F and C

Once that step is complete, you have your final string where the items in the second and fifth index positions are swapped:

String ABFDECGH

We just re-created the steps of a swap operation by using a temporary variable and some simple string shenanigans. The code for doing all of this is pretty straightforward, so let's quickly look at that next.

You can do a swap with any two values in an string using the [ ] operator.  Just put the index number in the square brackets (remember: index numbers always start with 0).  Here is an example of the code:
A for loop is a repetition control structure that allows you to efficiently write a loop that needs to execute a specific number of times.

A for loop is useful when you know how many times a task is to be repeated.

The for loop decision tree

Syntax:
The syntax of a for loop is:

for(initialization; Boolean_expression; update)
{
   //Statements
}
Here is the flow of control in a for loop:

Where to start the loop: The initialization step is executed first, and only once. This step allows you to declare and initialize any loop control variables. 

Whether to keep the loop going: Next, the Boolean expression is evaluated. If it is true, the loop keeps looping.  When it is true, the body of the loop is executed. If it is false, the loop stops. When it is false, the body of the loop does not execute and flow of control jumps to the next statement past the for loop.

How to make the Boolean expression false: After the body of the for loop executes, the flow of control jumps back up to the update statement. This statement allows you to update any loop control variables. The goal is to make progress toward stopping the loop, which happens when the Boolean condition is false.  So, this step is critical.  You must make it to false to stop the loop.

The Boolean expression is now evaluated again. If it is true, the loop executes and the process repeats itself (body of loop, then update step,then Boolean expression). After the Boolean expression is false, the for loop terminates.

Example: