Create The Modulo Operator %
If you combine the first two bullet points, you come up with this interesting feature of the modulo operator:  Whatever you put on the right side of the modulo operator will be one more than the upper limit of the result.  So, if that number is 2, the result will be in a range between 0 - 1.  If that number is 9, the result will be in a range between 0-8.  And so on.

The modulo operator is an extremely useful little operator because of that second bullet point: it limits the range of a number.  When working with computer memory, having a tool to limit the range of a number (aside from choosing a smaller data type) is invaluable.  We will see this when working with strings and arrays.

So want to try it out?  Let's look at the earlier example again and run it with all kinds of numbers instead of the number 9.  We will use a while loop to divide any number by 2 (up to the int data type limit).  Notice that the modulo output is limited to 0 to x-1, where the number on the right side of the modulo operator is x (divisor).  Since we use the number 2 in the x position, this means that every output will either be 0 or 1 for the following example:

int main() {

int modtest = 1;
int result = 0;
cout << "Enter a number: (q to quit): ";
  while(cin >> modtest)
  {
    result = modtest % 2;
    cout << result << endl;
    cout << "Enter a number: (q to quit): ";
  } 
  cout << "Thanks for playing!";
}
Testing modulo with a while loop 

There are a couple of interesting things about the while loop in this example:

The conditional statement of the while loop isn't actually a condition.  It is a cin statement.  This works because cin will return true if the operation succeeds in storing the input in the variable and it will return false if it fails.  Conditional statements require a true or false so it works.  We usually construct a conditional statement to get to a true or false, but that is optional.  All you really need is a boolean.  So, to keep the loop going, you need cin to succeed.  To stop the loop, you need cin to fail.

The loop is controlled by the input variable modtest, which is defined as an int.  However, to exit the loop, the user must type a letter (q).  This is because any non-valid number for modtest will cause the program to crash.  Try it with a 0 or a negative number and you will see what I mean.  We will see why this happens when we get to modulo operator restrictions in the next section. 

The only other option is to use a non-number.  The good news is that when you input a non-number with cin, your program doesn't crash.  The bad news is that it makes cin crash, which in this case isn't so bad because we don't plan to use cin anywhere else in the program and when cin crashes, it fails and returns false.  That is exactly what we need to end the loop. It's a little bit of a hack, but it works out.  

Restrictions of the modulo operator: 
The modulo operator has some restrictions or limitations.

The % operator cannot be applied to floating-point numbersLinks to an external site. i.e float or double. If you try to use the modulo operator with floating-point constants or variables, the compiler will produce a error:
// Program to illustrate the
// working of modulo operator
#include 
using namespace std;
 
int main()
{
     
    // To store two integer values
    float x, y;