Create loopEXAMPLES
hy is the while loop outside the brackets in the function : do {
    std::cout << "Enter A,B,C or D for question " << k << ":";
    std::cin >> answer;
    summary = summary + " " + answer; // add k onto summary
    k++;
 }while( k <= 10);  //10 question quiz
View other drafts



















The while loop is outside the brackets in the function because it is used to control the flow of the entire function. The function is a quiz that asks the user 10 questions. The while loop ensures that the function will loop through all 10 questions, even if the user gets some of them wrong.

If the while loop were inside the brackets, then the function would only ask the user a single question and then return. This is because the while loop would only be executed once, after the function is called.

By placing the while loop outside the brackets, the function will continue to ask the user questions until they have answered all 10 questions.

Here is a diagram that shows how the function works:

quiz()
{
  summary = "";
  k = 1;

  // While loop outside the brackets controls the flow of the entire function.
  while (k <= 10)
  {
    // Ask the user a question.
    std::cout << "Enter A,B,C or D for question " << k << ":";
    std::cin >> answer;

    // Add the user's answer to the summary.
    summary = summary + " " + answer;

    // Increment the question counter.
    k++;
  }

  // Return the summary of the user's answers.
  return summary;
}
This is a common pattern used in functions that need to perform a series of tasks repeatedly, such as reading a file line by line or processing a list of items.