Create  indefinite counting loop cycles an unknown number of times
why does not n print out multiple times in the function #include <string>
#include <iostream>
using namespace std;
int main() {
int n =9, count;

count = 0;
while ( n > 1 )
{
 n /= 2;
 count++;
}
std::cout << count << " time(s)"<< endl ;
std::cout << n << endl;
}
View other drafts


















The reason why n does not print out multiple times in the function is because the variable count is used to keep track of the number of times the loop iterates. The loop iterates while n is greater than 1. On each iteration, n is divided by 2 and count is incremented. Once n is less than or equal to 1, the loop terminates.

The variable n is only printed out once, after the loop has terminated. At this point, n is equal to 0.

Here is a breakdown of the function:

C++
int main() {
  int n = 9, count;

  count = 0;
  while (n > 1) {
    n /= 2;
    count++;
  }

  std::cout << count << " time(s)" << endl;
  std::cout << n << endl;
}
Use code with caution. Learn more
Line 1: The variable n is declared and initialized to the value 9. The variable count is declared and initialized to the value 0.
Line 3: The loop starts.
Line 4: The variable n is divided by 2.
Line 5: The variable count is incremented.
Line 6: The loop condition is checked. If n is greater than 1, the loop iterates again. Otherwise, the loop terminates.
Line 7: The variable count is printed out, followed by the string " time(s)".
Line 8: The variable n is printed out.
Since the variable n is only printed out once, after the loop has terminated, it is not printed out multiple times.

he while loop in your code will divide n by 2 until n is less than or equal to 1. Since n is initially set to 9, the loop will iterate 4 times:

Iteration 1: n = 9 / 2 = 4
Iteration 2: n = 4 / 2 = 2
Iteration 3: n = 2 / 2 = 1
Iteration 4: n = 1 / 2 = 0.5
The loop will then terminate because n is now less than or equal to 1.

However, it is important to note that the division operator in C++ is integer division. This means that the result of the division is always truncated to the nearest integer. So, in the last iteration of the loop, the result of the division is actually 0, not 0.5.

Therefore, the final value of n will be 0.

Here is a C++ code example:

C++
int main() {
  int n = 9;

  while (n > 1) {
    n /= 2;
  }

  std::cout << n << std::endl;

  return 0;
}
Use code with caution. Learn more