Create calcPCT() Math with a Loop
how does assigning pct to calcPCT() allow the arguments QuizTotal and WrongAnswers
to be called in calcPCT substituting for the arguments total and lostpoints in : #include <string>
#include <iostream>
using namespace std;
float calcPCT(float total, float lostpoints)
{
 float pct;
 pct = (total - lostpoints)/total;
 return pct;
}
int main()
{
 float QuizTotal = 0;
 float WrongAnswers;
 float pct;

do {
 cout << "Enter total possible points for the Quiz (greater than zero): ";
 cin >> QuizTotal;
} while (QuizTotal < 0);
 cout << " Please enter the number of points you missed on the quiz: ";
 cin >> WrongAnswers;
while (WrongAnswers >= QuizTotal)
{
 cout << "Error: missed points cannot be higher than total points. \n";
 cout << "Please enter the number of points you missed on the quiz: ";
 cin >> WrongAnswers;
}
pct = calcPCT(QuizTotal, WrongAnswers);
cout << pct << endl;
}

When you assign the result of calcPCT() to the variable pct, the arguments QuizTotal and WrongAnswers are automatically passed to the function calcPCT(). This is because the function is being called as a value, meaning that the return value of the function is being assigned to the variable pct.

When a function is called as a value, the arguments are passed to the function by copy. This means that the function does not have access to the original variables, but only to copies of them. This is important to keep in mind, because it means that any changes made to the arguments inside the function will not be reflected in the original variables.

In the example you provided, the function calcPCT() calculates the percentage of correct answers on a quiz. The arguments total and lostpoints represent the total number of possible points on the quiz and the number of points that were lost, respectively.

When the function is called from the main() function, the arguments QuizTotal and WrongAnswers are passed to the function by copy. The function then calculates the percentage of correct answers and returns the result. The result is then assigned to the variable pct in the main() function.

Here is a breakdown of the code:

C++
float pct = calcPCT(QuizTotal, WrongAnswers);
Use code with caution. Learn more
The function calcPCT() is called with the arguments QuizTotal and WrongAnswers.
The function calculates the percentage of correct answers and returns the result.
The result is assigned to the variable pct in the main() function.
Once the function has been called and the result has been assigned to the variable pct, the variable pct can be used as needed. For example, it could be printed to the console, displayed in a GUI, or used in other calculations.

profile picture
what does it mean to call a function as a value
View other drafts























Calling a function as a value means that the return value of the function is assigned to a variable. When a function is called as a value, the arguments are passed to the function by copy. This means that the function does not have access to the original variables, but only to copies of them. This is important to keep in mind, because it means that any changes made to the arguments inside the function will not be reflected in the original variables.

Here is an example of calling a function as a value:

C++
int add_one(int x) {
  return x + 1;
}

int main() {
  int y = add_one(10); // This is calling the function as a value.
  std::cout << y << std::endl; // This will print 11 to the console.
  return 0;
}
Use code with caution. Learn more
In this example, the function add_one() takes an integer as an argument and returns the argument plus one. The function is called from the main() function with the argument 10. The return value of the function is assigned to the variable y. The variable y is then printed to the console, which prints the value 11.

Calling functions as values can be useful in a number of situations. For example, it can be used to pass the results of one function to another function. It can also be used to create new variables that contain the results of function calls.

Here are some of the benefits of calling functions as values:

It can help to improve the readability of your code.
It can help to make your code more modular and reusable.
It can help to prevent side effects.
However, there are also some potential drawbacks to calling functions as values:

It can make your code less efficient, because the function arguments are copied when the function is called.
It can make your code more difficult to debug, because it can be difficult to track down the source of side effects.
Overall, whether or not to call a function as a value is a decision that should be made on a case-by-case basis. You should weigh the benefits and drawbacks of each approach and choose the approach that is best for your specific situation.