Create time clock controller
why is the std::cout<< "The target time is: " printed out
before the std::cout << counter from the void countDown() function in: #include <string>
#include <iostream>
using namespace std;
void countDown()
{
 int counter = 10;
 while (counter > 0)
 { std::cout << counter << std::endl;
  counter--;
}  
 std::cout << "Goodbye to now";
}

void displayTime(int hour, int minute)
{ 
 std::cout << "The target time is: " << hour;
 std::cout << ":" << minute << std::endl;
}
View other drafts


















The std::cout << "The target time is: " statement is printed out before the std::cout << counter from the void countDown() function in the code you provided because the displayTime() function is called before the countDown() function is called.

In C++, the order in which statements are executed is determined by the flow of the program. The flow of the program is controlled by statements such as if, else, while, and for.

In the code you provided, the displayTime() function is called before the countDown() function is called. This means that the statements inside the displayTime() function will be executed before the statements inside the countDown() function are executed.

The std::cout << "The target time is: " statement is inside the displayTime() function. This means that it will be executed before the std::cout << counter from the void countDown() function is executed.

Here is a diagram that shows the flow of the program:

void main() {
  displayTime(10, 30);
  countDown();
}

void displayTime(int hour, int minute) {
  std::cout << "The target time is: " << hour;
  std::cout << ":" << minute << std::endl;
}

void countDown() {
  int counter = 10;
  while (counter > 0) {
    std::cout << counter << std::endl;
    counter--;
  }
  std::cout << "Goodbye to now";
}
As you can see, the displayTime() function is called before the countDown() function is called. This means that the std::cout << "The target time is: " statement will be executed before the std::cout << counter from the void countDown() function is executed.