Create The Truth Table of Logical Operations
No matter how extensive a logical expression, all boil down to a binary true or false value when evaluated.

Considering only the result of an operand “a,” an operand “b” and the logical operator, we can construct the following table that shows the output of a given logical operation. 

a	b	a && b	a || b	  !a
true	true	true	true	false
true	false	false	true	false
false	false	false	false	true
false	true	false	true	true
This table provides a good way to check your work with logical operations. A program that does not return the listed result for the criteria provided will have a mistake that needs resolution.