Create LOOPCONTROLSTRATEGIES
Before you decide which loop to choose, you need to decide how to control it.  We will explore two possible ways to control a loop:

Counter Controlled Loop
User Controlled Loop
A counter controlled loop uses a variable that counts the number of times a loop must execute.  Use this strategy if you know exactly how many times you need to repeat your loop body.    The variable is set to a start value and increments each time through the loop until the variable value reaches the desired number of iterations.

A user controlled loop uses a variable that is set by the user through the user interface logic of your program.  When the user inputs a value that makes the loop condition false, the loop ends. Until the user provides that value, the loop continues. For example, the program might display something like "Press 0 to exit" or "Do you wish to continue (Y/N)?".

Note that you can use any loop construct to control your loop with either the counter controlled loop or user controlled loop, but some loops are better suited for certain control strategies.  The for loop is a great choice for a counter controlled loop because it has a counter variable built-in and the while loop is a better choice for a user controlled loop precisely because there is no counter variable built-in.  

The Loop Control
A loop does a statement (or block of statements) zero or many times, depending on the values controlling it. The part of the statement that manages the number of iterations of the loop is called the loop control. The statement, or block of statements, that is repetitively executed is called the loop body. The general form is

   <loop control>
      <loop body> ;

If the loop body contains more than one statement, you must enclose all the statements in braces like so:

   <loop control>
   {
      <body statement>;
      <body statement>;
      <body statement>;
      <body statement>;
      <body statement>;
   }
It is very important to indent the loop body. The compiler doesn't care, but we humans do. It makes it easy to see which and how many statements of the loop body are controlled by the loop control.

Counter Controlled Loop Example: Simple For Loops
If we wanted to count to 10, printing each number as we went, a for-loop would do the job:

   int k;
   
   for (k = 1; k <= 10; k = k + 1)
      cout << "k = " << k;
The explanation of this loop is what you might expect,but we'll analyze it carefully.

The top line is called the loop control. The second (indented) line is called the the loop body or loop block. The loop body is the thing that will be repeated. Each individual execution of the loop body is called a loop pass, or just pass.

The loop control tells us the following:
It describes how to begin the loop. This is called the init or initializer of the loop. The first part, k = 1, says start the loop by setting k to 1.
It explains how far to loop. This is called the test. The second part says keep looping as long as k <= 10, and if it ever becomes false, end the loop), and
It describes what to do at the end of each loop pass. This is called the iteration. k = k + 1 says to add 1 to k at the end of each pass).
When the loop ends, the program executes the next statement after the loop body. 

Guess, then confirm by running for yourself, what the output of the above loop looks like.

This for loop is good if you know, beforehand, how many times something will be repeated.  If the number of times that something is to be repeated is determined based on user input or other factors known only at run time, then a while loop is preferred.

For example, if I had said, "add all the odd numbers from 15 to 1001," then you should be able to do it simply with a loop rather than having one of those really long statements like 15 + 17 + 19 + ...  (sigh)  ... + 999 + 1001.  Give it a shot on your own.   Then compare with my solution:

#include <iostream>
using namespace std;

int  main()
{
   // make these symbolic constants so that if we wish to change them
   // we can do it quickly up here rather than muck around in the code
   int const LOW_VAL = 15;
   int const HIGH_VAL = 1001;

   int sum;    // holds the answer
   int k;      // loop counter

   sum = 0;    // start it off at 0 - we will add to it
   // loop, adding the ODD numbers
   for ( k = LOW_VAL; k <= HIGH_VAL; k = k + 2 )
      sum = sum + k;
 
   cout << "The sum of the odd numbers between " << LOW_VAL
      << " and " << HIGH_VAL << ", inclusive, is " << sum << endl << endl;

   return 0;
}

/* ------------------- Sample Run --------------------

Run the program to get the answer.
User Controlled Loop Example: While Loops
Simpler than for loops, while loops are used when you don't have so much bookkeeping to do. The for loop had three control sections:  init, test, iteration.  The while loop only has one, the test. The idea is that you test the condition in the while statement.  If it is true, you execute the loop body, then test again after the loop pass.  As long as the condition remains true, you keep on looping.  If it ever becomes false, you end the loop and move on to the statement following the loop body. 

In the following excerpt from a program we ask the user for a value.  If they give us a number less than 20 we slap their wrists and ask again.  When they finally follow instructions, we leave them alone and end the input loop:

   double withdrawal;

   // get the withdrawal amount from user
   cout << "Amount of withdrawal ($20 minimum): ";
   cin >> withdrawal;
      
   // trap user until they follow directions
   while (withdrawal < 20)
   {
      cout << "$20 Minimum.  \nPlease try again: ";
      cin >> withdrawal;
   }
   
/* -------------------

Amount of withdrawal ($20 minimum): 18
$20 Minimum.  
Please try again: 25

New balance:975.03

... etc.

------------------- */
This example was obviously not a complete program but merely a code fragment.  It demonstrates a while loop and also gives a reasonable input filter on which you can base future assignments.  An input filter is a chunk of code that traps the user until he follows directions.  There is no other way out.  You may want to give the user a chance to quit trying, which will require a secondary condition.  We will explore multiple conditions in a loop in the next topic.